Don't delete files
Previously the Upstart (or other) files were unlinked before re-writing
them. It's not clear why this was done, as writing to an existing file
with the default flags truncates the file which has the same overall
effect.

Additionally, the pattern used to determining which files to delete was
overly broad and willing to delete important files that happened to
have the app's name as a prefix.

Closes #55