Adds support for SSL proxy
If you have server key and cert files, then you can have your proxy listen over
https as well as http.  This is very handy for when your app has features that
require SSL to work, such as accessing WebRTC streams.

An SSL proxy is started when the --ssl-key and --ssl-cert options are set with
valid key and certificate files.  To run locally, it may be necessary to create
a self-signed certificate.